+++ packages/realtime/package.json
@@ -1,6 +1,8 @@
 {
   "name": "@u22n/realtime",
   "version": "0.0.0",
+  "private": true,
+  "types": "./index.ts",
   "main": "index.js",
   "scripts": {
     "check": "tsc --noEmit"
